+++ src/main/java/org/visuals/legacy/animatium/mixins/v1/rendering/items/MixinItemInHandRenderer.java
@@ -186,15 +186,15 @@ public abstract class MixinItemInHandRenderer {
 					from = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/ItemStack;getUseAnimation()Lnet/minecraft/world/item/ItemUseAnimation;"),
 					to = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/ItemInHandRenderer;applyItemArmTransform(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/world/entity/HumanoidArm;F)V", ordinal = 6)
 			))
-	private void animatium$itemUsageSwinging(final AbstractClientPlayer player, final float partialTick, final float pitch, final InteractionHand hand, final float swingProgress, final ItemStack item, final float equippedProgress, final PoseStack poseStack, final SubmitNodeCollector nodeCollector, final int packedLight, final CallbackInfo ci, @Local(name = "humanoidArm") HumanoidArm arm) {
+	private void animatium$itemUsageSwinging(final AbstractClientPlayer player, final float partialTick, final float pitch, final InteractionHand hand, final float swingProgress, final ItemStack item, final float equippedProgress, final PoseStack poseStack, final SubmitNodeCollector nodeCollector, final int packedLight, final CallbackInfo ci, @Local HumanoidArm arm) {
 		if (Animatium.isEnabled() && AnimatiumConfig.instance().items.itemUsageSwinging) {
-			applyItemArmAttackTransform(poseStack, arm, swingProgress);
+			this.applyItemArmAttackTransform(poseStack, arm, swingProgress);
 		}
 	}
 
 	@WrapOperation(method = "tick", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/player/LocalPlayer;getAttackStrengthScale(F)F"))
 	private float animatium$highAttackSpeedVisual(final LocalPlayer instance, final float defaultAttackScale, final Operation<Float> original) {
-		if (Animatium.isEnabled() && AnimatiumConfig.instance().other.highAttackSpeedVisual) {
+		if (Animatium.isEnabled() && AnimatiumConfig.instance().extras.highAttackSpeedVisual) {
 			return defaultAttackScale;
 		} else {
 			return original.call(instance, defaultAttackScale);