+++ src/main/java/org/visuals/legacy/animatium/util/rendering/LegacyLogoTexture.java
@@ -39,15 +39,15 @@ import java.io.InputStream;
 
 // LoadingOverlay#LogoTexture but customizable
 public class LegacyLogoTexture extends ReloadableTexture {
-    public LegacyLogoTexture(ResourceLocation resourceLocation) {
-        super(resourceLocation);
-    }
+	public LegacyLogoTexture(ResourceLocation resourceLocation) {
+		super(resourceLocation);
+	}
 
-    @Override
-    public @NotNull TextureContents loadContents(ResourceManager resourceManager) throws IOException {
-        try (final InputStream inputStream = Minecraft.getInstance().getResourceManager().open(this.resourceId())) {
-            // TODO: Get real metadata file
-            return new TextureContents(NativeImage.read(inputStream), new TextureMetadataSection(TextureMetadataSection.DEFAULT_BLUR, TextureMetadataSection.DEFAULT_CLAMP));
-        }
-    }
+	@Override
+	public @NotNull TextureContents loadContents(ResourceManager resourceManager) throws IOException {
+		try (final InputStream inputStream = Minecraft.getInstance().getResourceManager().open(this.resourceId())) {
+			// TODO: Get real metadata file
+			return new TextureContents(NativeImage.read(inputStream), new TextureMetadataSection(TextureMetadataSection.DEFAULT_BLUR, TextureMetadataSection.DEFAULT_CLAMP));
+		}
+	}
 }