+++ CHANGELOG.md
@@ -8,13 +8,24 @@
   
 #### General
 
+- Old Crafting Slots Position
+- Item Swing Settings
+- Old Water Color Effects
 - Fix thrown egg kicking you
+- Option to turn on/off 1.7 "longer un-sneak" animation
+- Singleplayer toggles for server features (in extras config tab)
 
 ## Changed:
 
 #### General
 
+- Fixed "Old Widget Text Colors" causing disabled buttons to have white text
+- Removed "Old Depth Far" & "Always Steve Model"
+- "Force High Attack Speed Visual" is now default to false & won't change based on preset
 - Refactored some settings to be reversed of what they were before (Missed in 3.0)
+- Changing settings no longer reloads resource pack unless the specific setting requiring it was changed
+- "Fast" settings are no longer dependent of video settings
+  - Presets will still set accordingly
 
 ### Server